requirements.txt
# Core runtime deps
pygame-ce
numpy
